+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.2] - 2020-06-10
+
 ## [4.0.1] - 2020-06-09
 ### Changed
 - Use the encoded resource path and query parameters for Mauth singer